+++ changelog.md
@@ -2,8 +2,13 @@
 
 ## Unreleased
 
+### Features
+
+- [#3830](https://github.com/ignite/cli/pull/3830) Remove gRPC info from Ignite Apps errors
+
 ### Fixes
 
+- [#3825](https://github.com/ignite/cli/pull/3825) Fix a minor Keplr type-checking bug in TS client
 - [#3833](https://github.com/ignite/cli/pull/3833) Improve Cosmos SDK detection to support SDK forks
 
 ## [`v28.0.0`](https://github.com/ignite/cli/releases/tag/v28.0.0)